downgrade rank to a short

+++ Tic-Stack-Toe/src/com/oakonell/ticstacktoe/model/rank/EloRanker.java
@@ -3,7 +3,7 @@ package com.oakonell.ticstacktoe.model.rank;
 public class EloRanker implements RankingRater {
 
 	@Override
-	public long calculateRank(long oldRank, long opponentsRank,
+	public short calculateRank(short oldRank, short opponentsRank,
 			GameOutcome outcome) {
 		double expected = calculateExpected(oldRank, opponentsRank);
 
@@ -20,10 +20,10 @@ public class EloRanker implements RankingRater {
 			break;
 		}
 		double newRank = oldRank + getKFactor(oldRank) * (actual - expected);
-		return (int) Math.round(newRank);
+		return (short) Math.round(newRank);
 	}
 
-	private double getKFactor(long oldRank) {
+	private double getKFactor(short oldRank) {
 		// https://en.wikipedia.org/wiki/Elo_rating_system
 		// Using USCF logistic distribution
 		// Players below 2100: K-factor of 32 used
@@ -36,13 +36,13 @@ public class EloRanker implements RankingRater {
 		return 16;
 	}
 
-	private double calculateExpected(long oldRank, long opponentsRank) {
+	private double calculateExpected(short oldRank, short opponentsRank) {
 		double denom = 1 + Math.pow(10, (opponentsRank - oldRank) / 400.0);
 		return 1.0 / denom;
 	}
 
 	@Override
-	public long initialRank() {
+	public short initialRank() {
 		return 1200;
 	}
